Final tuning for the particle search

+++ SOURCE_FILES/IVP_SOLVER.cu
@@ -174,7 +174,7 @@ auto end = std::chrono::high_resolution_clock::now();
 	const uint64_t MEM_2_STORE = store_PDFs.size() * sizeof(float);
 	
 	uint32_t number_of_frames_needed 	= MEM_2_STORE / Grid_Nodes / sizeof(float);
-	uint32_t max_frames_file 			= MAX_FILE_SIZE_B / Grid_Nodes / sizeof(float);
+	uint64_t max_frames_file 			= MAX_FILE_SIZE_B / Grid_Nodes / sizeof(float);
 	uint32_t number_of_files_needed  	= floor((number_of_frames_needed - 1) / max_frames_file) + 1;
 	
 	char ans;
